removed ARMV6 support, added ARM64 support

+++ MapView/Map/RMMapOverlayView.h
@@ -12,7 +12,7 @@
 
 @interface RMMapOverlayView : UIView
 
-- (unsigned)sublayersCount;
+- (NSUInteger)sublayersCount;
 
 - (void)addSublayer:(CALayer *)aLayer;
 - (void)insertSublayer:(CALayer *)aLayer atIndex:(unsigned)index;
