+++ MapView/Map/RMMapOverlayView.m
@@ -28,7 +28,7 @@
     return self;
 }
 
-- (unsigned)sublayersCount
+- (NSUInteger)sublayersCount
 {
     return [self.layer.sublayers count];
 }
