+++ MapView/Map/RMMapView.m
@@ -949,7 +949,7 @@
 
     _mapScrollViewIsZooming = NO;
 
-    int tileSideLength = [_tileSourcesContainer tileSideLength];
+    NSUInteger tileSideLength = [_tileSourcesContainer tileSideLength];
     CGSize contentSize = CGSizeMake(tileSideLength, tileSideLength); // zoom level 1
 
     _mapScrollView = [[RMMapScrollView alloc] initWithFrame:[self bounds]];
@@ -1099,7 +1099,7 @@
         [_delegate beforeMapZoom:self];
 }
 
-- (void)scrollViewDidEndZooming:(UIScrollView *)scrollView withView:(UIView *)view atScale:(float)scale
+- (void)scrollViewDidEndZooming:(UIScrollView *)scrollView withView:(UIView *)view atScale:(CGFloat)scale
 {
     _mapScrollViewIsZooming = NO;
 
@@ -1690,7 +1690,7 @@
     }
     else
     {
-        int tileSideLength = [_tileSourcesContainer tileSideLength];
+        NSUInteger tileSideLength = [_tileSourcesContainer tileSideLength];
         CGSize contentSize = CGSizeMake(tileSideLength, tileSideLength); // zoom level 1
 
         RMMapTiledLayerView *tiledLayerView = [[RMMapTiledLayerView alloc] initWithFrame:CGRectMake(0.0, 0.0, contentSize.width, contentSize.height) mapView:self forTileSource:newTileSource];
