+++ MapView/Map/RMTileSourcesContainer.m
@@ -39,7 +39,7 @@
     RMSphericalTrapezium _latitudeLongitudeBoundingBox;
 
     float _minZoom, _maxZoom;
-    int _tileSideLength;
+    NSUInteger _tileSideLength;
 }
 
 - (id)init
