+++ Proj4/pj_open_lib.c
@@ -117,7 +117,7 @@ pj_open_lib(projCtx ctx, char *name, char *mode) {
     if (*name == '~' && strchr(dir_chars,name[1]) )
         if ((sysname = getenv("HOME")) != NULL) {
             (void)strcpy(fname, sysname);
-            fname[n = strlen(fname)] = DIR_CHAR;
+            fname[n = (int)strlen(fname)] = DIR_CHAR;
             fname[++n] = '\0';
             (void)strcpy(fname+n, name + 1);
             sysname = fname;
@@ -138,7 +138,7 @@ pj_open_lib(projCtx ctx, char *name, char *mode) {
     /* or is environment PROJ_LIB defined */
     else if ((sysname = getenv("PROJ_LIB")) || (sysname = proj_lib_name)) {
         (void)strcpy(fname, sysname);
-        fname[n = strlen(fname)] = DIR_CHAR;
+        fname[n = (int)strlen(fname)] = DIR_CHAR;
         fname[++n] = '\0';
         (void)strcpy(fname+n, name);
         sysname = fname;
